Fix to ResampleSfx() function which, in both the original and the floating-point processor, was not performing correctly the Catmull-Rom interpolation for supersampled sound effects due to an incorrectly typed source sample fraction variable in code. Performed clamping of the new samples into the range for short (signed 16-bit) values, to avoid some audio artifacts after applying the previous correction.  Minor code indent adjustments performed as well.

+++ SlipNFrag-XR/app/src/main/cpp/snd_mem_oxr.cpp
@@ -11,11 +11,11 @@ void ResampleSfx (sfx_t *sfx, int inrate, int inwidth, byte *data)
 {
 	int		outcount;
 	int		srcsample;
-    int     srcsamplefrac;
+	float	srcsamplefrac;
 	float	stepscale;
 	int		i;
 	int		y0, y1, y2, y3, samplefrac, fracstep;
-    float   a0, a1, a2, a3, mu, mu2;
+    float	a0, a1, a2, a3, mu, mu2;
 	sfxcache_t	*sc;
 	
 	sc = (sfxcache_t*)sfx->data;
